+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-# 0.0.1
+# 0.1.0 (unreleased)
+
+* add ability to refer to a revision of a git repo
+* improve diff color UX
+* added `-k` / `--keep` to skip replacement prompts and keep local copies of files
+* added `-y` / `--yes` to skip prompts and accept all default values
+
+# 0.0.2
 
 * fix issue with local repository uris
 * add support for providing a custom root directory within a template repository